🟥🟩🟦 Si_ExisteUnKiloDeManzanaEnElCarritoYSeAgregoUnaOfertaDelVeintePorCiento_Debe_GenerarElReciboConProductoYValorDescuentoAplicado

+++ KatasTDD.Domain/Supermercado/DTO/Oferta.cs
@@ -1,10 +1,10 @@
 ﻿namespace KatasTDD.Domain.Supermercado.DTO;
 
-public class Oferta(TipoOferta tipo, Producto producto, double? valorOferta = null)
+public class Oferta(TipoOferta tipo, Producto producto, decimal? valorOferta = null)
 {
     public TipoOferta Tipo { get; } = tipo;
     public Producto ProductoAplicado { get; } = producto;
-    public double? ValorOferta { get; } =  valorOferta;
+    public decimal? ValorOferta { get; } =  valorOferta;
 }
 
 public enum TipoOferta
